PAE-1536: extract validation issue logging into its own module
